% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCGPier.R
\name{PCGPier}
\alias{PCGPier}
\title{Calculate the gene-pair based prognostic signature PCGPier for prostate cancer samples}
\usage{
PCGPier(expr = NULL, res.path = getwd())
}
\arguments{
\item{expr}{A numerical expression matrix or data frame with row for gene symbol name and column for sample ID. Note: In principle, the expression profile does not need any normalization, but since the amount of gene expression is affected by the gene length, the original count or the normalized count data may not be suitable for this analysis. It is recommended to provide FPKM or TPM value and log transformation is not necessary.}

\item{res.path}{A string value to indicate the output path for storing the Estimated IRGPI.txt file.}
}
\value{
A data frame stored sample name and estimated IRGPI.
}
\description{
This prognosis prediction study established a gene-pairs based RFS prediction classifier for prostate cancer patients, which is universal applicable in multiple cohorts with different gene expression detection methods.
}
\examples{
library(PCGPier)
load(system.file("extdata", "demo.RData", package = "PCGPier", mustWork = TRUE)) # load example data
riskscore  <- PCGPier(expr = demo)
}
\author{
Jialin Meng, Xiaofan Lu, Tong Jin
}
